+++ Authorization/Creatable/Base.ts
@@ -18,12 +18,12 @@ export namespace Base {
 		return typeof value == "object" &&
 			(value.number == undefined || typeof value.number == "string") &&
 			(value.descriptor == undefined || typeof value.descriptor == "string") &&
-			(value.ip == undefined || typeof value.ip == "string") && (
-				typeof value.amount == "number" &&
-				isoly.Currency.is(value.currency) &&
-				(value.account == "create" || value.account == undefined || authly.Token.is(value.account)) ||
-				value.account == "create" && value.amount == undefined && value.currency == undefined
+			(value.ip == undefined || typeof value.ip == "string") &&
+			(
+				value.amount == undefined && value.currency == undefined ||
+				typeof(value.amount) == "number" && isoly.Currency.is(value.currency)
 			) &&
+			(value.account == "create" || value.account == undefined || authly.Token.is(value.account)) &&
 			(value.pares == undefined || typeof value.pares == "string") &&
 			(value.callback == undefined || typeof(value.callback) == "string")
 	}
